Remove direct users of TypedMethod.execute
Instead, pull functionality shared across multiple API methods into
non-method helper functions.
