+++ imports/server/methods/rollAPIKey.ts
@@ -1,7 +1,7 @@
 import { check, Match } from 'meteor/check';
 import Logger from '../../Logger';
-import fetchAPIKey from '../../methods/fetchAPIKey';
 import rollAPIKey from '../../methods/rollAPIKey';
+import ensureAPIKey from '../ensureAPIKey';
 import APIKeys from '../models/APIKeys';
 import userForKeyOperation from '../userForKeyOperation';
 
@@ -22,6 +22,6 @@ rollAPIKey.define({
       await APIKeys.destroyAsync(k._id);
     }
 
-    return fetchAPIKey.execute(this, { forUser });
+    return ensureAPIKey({ requestedBy: this.userId, forUser });
   },
 });
